// src/lib/src/shared/flo.common.d.ts
export declare namespace Flo {
    enum DnDEventType {
        DRAG = 0,
        DROP = 1,
    }
    interface DnDEvent {
        type: DnDEventType;
        view: dia.CellView;
        event: MouseEvent;
    }
    interface PropertyMetadata {
        readonly id: string;
        readonly name: string;
        readonly description?: string;
        readonly defaultValue?: any;
        readonly [propName: string]: any;
    }
    interface ExtraMetadata {
        readonly titleProperty: string;
        readonly noEditableProps: boolean;
        readonly noPaletteEntry: boolean;
        readonly [propName: string]: any;
        readonly allowAdditionalProperties: boolean;
    }
    interface ElementMetadata {
        readonly name: string;
        readonly group: string;
        description?(): Promise<string>;
        get(property: String): Promise<PropertyMetadata>;
        properties(): Promise<Array<PropertyMetadata>>;
        readonly metadata?: ExtraMetadata;
        readonly [propName: string]: any;
    }
// ...
}